+++ src/renderer/components/Dashboard/Dashboard.tsx
@@ -3,6 +3,7 @@ import { Globe } from '../Globe/Globe';
 import { HudStatusPanel } from '../HudStatusPanel/HudStatusPanel';
 import { HudCollectorPanel } from '../HudCollectorPanel/HudCollectorPanel';
 import { HudEventPanel } from '../HudEventPanel/HudEventPanel';
+import { LeftColumn, RightColumn } from '../SideColumns';
 import { SkyMapModal } from '../SkyMapModal/SkyMapModal';
 import { SourceDirectoryModal } from '../SourceDirectoryModal/SourceDirectoryModal';
 import { Ticker } from '../Ticker/Ticker';
@@ -37,6 +38,10 @@ export function Dashboard() {
 
       {/* Top-right: collector badges + sky map toggle */}
       <HudCollectorPanel />
+
+      {/* Left/right side columns: mission-control telemetry widgets, z-10 */}
+      <LeftColumn />
+      <RightColumn />
 
       {/* Right center: event detail slide-in */}
       <HudEventPanel />